requirements.txt
google-genai>=1.0.0
edge-tts>=6.1.0
asyncio
argparse